/GET radius and polygon
In order to stay Rubocop compliant GpsController was split into 2 new modules
- GpsControllerCreate manage the create action
- GpsControllerIndex manage the index action

Since there is no defined semantic for GET parameters, we are also using POST method
to retreive the points within radius and polygon and use the header `X-HTTP-Method-Override`
with the value `GET` to make the web server think we are indeed answering a GET request.
This allow us to have the same semantic for both POST and GET requests.

The Repository layer is responsible to speak to the database and return ready to be used datasets.